+++ NEWS.md
@@ -95,6 +95,7 @@
 * Add a variety of indirect response models
 * Add a variety of tumor growth inhibition models and move all oncology models
   into a new model database directory
+* Add a variety of double-absorption PK models
 * `cp` and related `cpddSd` and `cppropSd` were renamed to `Cc`, `CcAddSd` and
   `CcPropSd` (fix #70).
 * Multiple-endpoint models will have the `DV` column in the modeldb separated by